+++ src/browser.ts
@@ -1,0 +1,2 @@
+export { getElementSourceLocation } from './getElementSourceLocation'
+