+++ src/server.ts
@@ -1,0 +1,2 @@
+export { resolveSourceLocationInServer } from './resolveSourceLocationInServer'
+